Moved Connection to its own file, did some refactoring in ads files (#269)
